+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/ConstructSpell.java
@@ -89,6 +89,8 @@ public class ConstructSpell extends Spell
 
 		int radius = parameters.getInt("radius", defaultRadius);		
 		String typeString = parameters.getString("type", "");
+		
+		radius = (int)(playerSpells.getPowerMultiplier() * radius);
 
 		ConstructionType testType = ConstructionType.parseString(typeString, ConstructionType.UNKNOWN);
 		if (testType != ConstructionType.UNKNOWN)