+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/TossSpell.java
@@ -34,8 +34,10 @@ public class TossSpell extends Spell
 		
 		int tossCount = 1;
 		tossCount = parameters.getInt("count", tossCount);
+		tossCount = (int)(playerSpells.getPowerMultiplier() * tossCount);	
 		float speed = 0.6f;
 		speed = (float)parameters.getDouble("speed", speed);
+		speed = (int)(playerSpells.getPowerMultiplier() * speed);	
 		
 		Vector direction = player.getLocation().getDirection();
 		direction.normalize().multiply(speed);